Add support for Arduino UNO R4 boards
The new UNO R4 boards are unlike previous UNO boards which
are 8 bit AVR based. but instead an ARM M4 board.  Their analogWrite code does not allow us to directly set the frequency of the PWM, but it is built on top of a PwmOut object which does.   I am using those objects.